requirements.txt
asgiref==3.2.7
Authlib==0.14.3
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.7
idna==2.9
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
xero-python==0.2.2
